bloomfilter: merge .Classic.Mutable into Classic.Internal and .Classic
The MBloom representation and functions that need access to it go into
Classic.Internal while functions that don't need representation access
go into the .Classic top level module.
