+++ lsm-tree.cabal
@@ -426,7 +426,6 @@ library bloomfilter
     Data.BloomFilter.Classic.BitVec64
     Data.BloomFilter.Classic.Calc
     Data.BloomFilter.Classic.Internal
-    Data.BloomFilter.Classic.Mutable
     Data.BloomFilter.Hash
 
   ghc-options:     -O2 -Wall
